Move DonateModal to modals
