+++ client/src/routes/modals/DonateModal/utils/currency.ts
